refactor: split admin settings and runtime services
