+++ gpt2giga/app/_admin_settings/__init__.py
@@ -1,0 +1,1 @@
+"""Internal building blocks for admin settings services."""
