+++ gpt2giga/features/batches/__init__.py
@@ -1,9 +1,9 @@
 """Batches capability."""
 
-from gpt2giga.features.batches.service import (
-    BatchesService,
-    get_batches_service_from_state,
-)
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
 from gpt2giga.features.batches.validation import (
     BatchInputValidator,
     detect_batch_input_format,
@@ -31,3 +31,22 @@ __all__ = [
     "validate_batch_input_bytes",
     "validate_batch_input_rows",
 ]
+
+if TYPE_CHECKING:
+    from gpt2giga.features.batches.service import (
+        BatchesService,
+        get_batches_service_from_state,
+    )
+
+
+def __getattr__(name: str) -> Any:
+    """Lazily expose the batches service surface."""
+    if name == "BatchesService":
+        from gpt2giga.features.batches.service import BatchesService
+
+        return BatchesService
+    if name == "get_batches_service_from_state":
+        from gpt2giga.features.batches.service import get_batches_service_from_state
+
+        return get_batches_service_from_state
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
